# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MdM Control Backend",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "List Devices",
        "operationId": "list_devices_api_v1_devices_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_DeviceResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Get Device",
        "operationId": "get_device_api_v1_devices__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Devices"
        ],
        "summary": "Update Device",
        "operationId": "update_device_api_v1_devices__device_id__put",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}/commands": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "List Device Commands",
        "operationId": "list_device_commands_api_v1_devices__device_id__commands_get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CommandResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Trigger Device Command",
        "operationId": "trigger_device_command_api_v1_devices__device_id__commands_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}/commands/{command_id}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "Get Device Command",
        "operationId": "get_device_command_api_v1_devices__device_id__commands__command_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          },
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Command Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "Cancel Device Command",
        "operationId": "cancel_device_command_api_v1_devices__device_id__commands__command_id__delete",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          },
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Command Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/extension-attributes": {
      "get": {
        "tags": [
          "Extension Attributes"
        ],
        "summary": "List Extension Attributes",
        "operationId": "list_extension_attributes_api_v1_extension_attributes_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ExtensionAttributeResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Extension Attributes"
        ],
        "summary": "Create Extension Attribute",
        "operationId": "create_extension_attribute_api_v1_extension_attributes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionAttributeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionAttributeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/extension-attributes/{attribute_id}": {
      "get": {
        "tags": [
          "Extension Attributes"
        ],
        "summary": "Get Extension Attribute",
        "operationId": "get_extension_attribute_api_v1_extension_attributes__attribute_id__get",
        "parameters": [
          {
            "name": "attribute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attribute Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionAttributeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Extension Attributes"
        ],
        "summary": "Update Extension Attribute",
        "operationId": "update_extension_attribute_api_v1_extension_attributes__attribute_id__put",
        "parameters": [
          {
            "name": "attribute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attribute Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionAttributeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionAttributeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Extension Attributes"
        ],
        "summary": "Delete Extension Attribute",
        "operationId": "delete_extension_attribute_api_v1_extension_attributes__attribute_id__delete",
        "parameters": [
          {
            "name": "attribute_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attribute Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandCreate": {
        "properties": {
          "command_type": {
            "$ref": "#/components/schemas/CommandType"
          }
        },
        "type": "object",
        "required": [
          "command_type"
        ],
        "title": "CommandCreate"
      },
      "CommandResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "device_id": {
            "type": "integer",
            "title": "Device Id"
          },
          "command_type": {
            "$ref": "#/components/schemas/CommandType"
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          },
          "result_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Message"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "command_type",
          "status",
          "created_at"
        ],
        "title": "CommandResponse"
      },
...
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "Online",
          "Offline",
          "Pending"
        ],
        "title": "ConnectionStatus"
      },
...
      "DeviceResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "serial_number": {
            "type": "string",
            "title": "Serial Number"
          },
          "os_version": {
            "type": "string",
            "title": "Os Version"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "enrollment_status": {
            "$ref": "#/components/schemas/EnrollmentStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "last_enrolled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Enrolled At"
          },
          "battery_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Status"
          },
          "total_storage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Storage"
          },
          "available_storage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Storage"
          },
          "total_memory": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Memory"
          },
          "available_memory": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Memory"
          },
          "network": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Network"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network"
          },
          "certificates": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Certificate"
                },
                "type": "array"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificates"
          },
          "extension_attributes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DeviceExtensionAttributeResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Attributes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "serial_number",
          "os_version",
          "connection_status",
          "enrollment_status",
          "created_at",
          "updated_at",
          "last_enrolled_at"
        ],
        "title": "DeviceResponse"
      },
      "DeviceUpdate": {
        "properties": {
          "connection_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "enrollment_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnrollmentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "battery_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Status"
          },
          "total_storage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Storage"
          },
          "available_storage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Storage"
          },
          "total_memory": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Memory"
          },
          "available_memory": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Memory"
          },
          "network": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Network"
              },
              {
                "type": "null"
              }
            ]
          },
          "certificates": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Certificate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificates"
          },
          "extension_attributes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DeviceExtensionAttributeCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extension Attributes"
          }
        },
        "type": "object",
        "title": "DeviceUpdate"
      },
...
      "ExtensionAttributeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "data_type": {
            "$ref": "#/components/schemas/ExtensionDataType"
          },
          "input_type": {
            "$ref": "#/components/schemas/ExtensionInputType"
          },
          "popup_choices": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Popup Choices"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "name",
          "data_type",
          "input_type"
        ],
        "title": "ExtensionAttributeCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_CommandResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommandResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PaginatedResponse[CommandResponse]"
      },
      "PaginatedResponse_DeviceResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DeviceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PaginatedResponse[DeviceResponse]"
      },
      "PaginatedResponse_ExtensionAttributeResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExtensionAttributeResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "PaginatedResponse[ExtensionAttributeResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}